% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genr.R
\name{genr}
\alias{genr}
\title{Title genr}
\usage{
genr(n = 100)
}
\arguments{
\item{n}{sample size generated}
}
\value{
generated data
}
\description{
Title genr
}
\examples{
generate<-genr()
}
\author{
Fadhaa Ali
}
