% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regress.R
\name{regress}
\alias{regress}
\title{This to apply lm model to data}
\source{
The data comes from simulation.
}
\usage{
regress(dat)
}
\arguments{
\item{dat}{The input dataframe}
}
\value{
summary of lm model
}
\description{
This is short description
}
\details{
This is package
}
\examples{
n<-100
x1<-rnorm(n,0,1);x2<-rnorm(n,0,1);Y<-2+3*x1+4*x2+rnorm(n,0,1)
dat<-data.frame(cbind(Y,x1,x2))
regress(dat)
}
\references{
Chambers, J. M. (1992) Linear models. Chapter 4 of Statistical Models in S eds J. M. Chambers and T. J. Hastie, Wadsworth & Brooks/Cole.

Wilkinson, G. N. and Rogers, C. E. (1973). Symbolic descriptions of factorial models for analysis of variance. Applied Statistics, 22, 392--399. 10.2307/2346786.
}
\author{
author Fadhaa Ali
}
\concept{Linear Regression.}
