% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/REGsimu-package.R
\docType{package}
\name{REGsimu-package}
\alias{REGsimu}
\alias{REGsimu-package}
\title{REGsimu: This package to simulate data and apply lm model}
\description{
This package is to apply lm model to input data and find the estimators values.
}
\keyword{internal}
