% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rg_data.R
\docType{data}
\name{rg_data}
\alias{rg_data}
\title{Data is generated via standard regression model}
\format{
data frame with 100 rows and number of variables

\describe{
    \item{the first col}{Y}
    \item{the second...}{X1...}
    }
}
\source{
{This dataset is created by simulation}
data(rg_data)
}
\usage{
rg_data
}
\description{
Contains the response variable Y and the independent variables $X_i$
}
\keyword{datasets}
