% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rgsim.R
\name{rgsim}
\alias{rgsim}
\title{This to generate datasets according to linear regression model}
\source{
The data comes from simulation.
}
\usage{
rgsim(n, nv)
}
\arguments{
\item{n}{The sample size}

\item{nv}{The number of desirable variables}
}
\value{
list of data and parameters used in simulation
}
\description{
This is short description
}
\details{
This is package
}
\examples{
n<-100
var<-3
dat<-rgsim(n,var)
print(dat$data)
print(dat$par)
}
\references{
some references
}
\author{
Author Fadhaa Ali
}
\concept{Linear Regression simulation.}
